Split DataSourceDashboard into components
